Fix digitalocean2 NumberFormatException (#216)
Changed org.jclouds.digitalocean2.domain.Action.id from int to long
 in order to fit large numbers sent in responses.

 This appears to be a misinterpretation of DigitalOceans ApiDocs,
 because in these docs, they fail to specify the size of their
 "integer" type. Therefore, I looked into their official GO Api
 sources. There, they use the int type, which is platform dependant
 and usually has 64 bits.

 Therefore our digitalocean2 provider, should be refactored, changing
 all Ids from int/Integer to long/Long.

+++ providers/digitalocean2/src/main/java/org/jclouds/digitalocean2/compute/extensions/DigitalOcean2ImageExtension.java
@@ -60,13 +60,13 @@ public class DigitalOcean2ImageExtension implements ImageExtension {
    protected Logger logger = Logger.NULL;
 
    private final DigitalOcean2Api api;
-   private final Predicate<Integer> imageAvailablePredicate;
+   private final Predicate<Long> imageAvailablePredicate;
    private final Predicate<Integer> nodeRunningPredicate;
    private final Function<ImageInRegion, Image> imageTransformer;
    private final ListeningExecutorService userExecutor;
 
    @Inject DigitalOcean2ImageExtension(DigitalOcean2Api api,
-         @Named(TIMEOUT_IMAGE_AVAILABLE) Predicate<Integer> imageAvailablePredicate,
+         @Named(TIMEOUT_IMAGE_AVAILABLE) Predicate<Long> imageAvailablePredicate,
          @Named(TIMEOUT_NODE_RUNNING) Predicate<Integer> nodeRunningPredicate,
          Function<ImageInRegion, Image> imageTransformer,
          @Named(Constants.PROPERTY_USER_THREADS) ListeningExecutorService userExecutor) {
